backend/requirements.txt
fastapi
uvicorn
neo4j
pandas
python-dotenv
faker